requirements.txt
asgiref==3.8.1
colorama==0.4.6
crispy-bootstrap4==2024.1
Django==5.1.1
django-cleanup==8.1.0
django-crispy-forms==2.3
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
